requirements.txt
streamlit==1.32.0
playwright==1.42.0
playwright-stealth==1.0.6
PyGithub==2.1.1
pandas==2.1.4
openai==1.14.0
